seperated out perft so that you can access it from ui
