+++ public/index.html
@@ -24,7 +24,8 @@
                 "./legal_move_generator": "./legal_move_generator.js",
                 "./pseudo_move_generator": "./pseudo_move_generator.js",
                 "./gui": "./gui.js",
-                "./move": "./move.js"
+                "./move": "./move.js",
+                "./perft": "./perft.js"
 			}
 		}
 	</script>
